Optimize e2e scripts
fix bug

add

fix shell
