+++ .github/workflows/auto-pr-ci.yaml
@@ -149,4 +149,4 @@ jobs:
           echo ${{ runner.os }}
           echo ${{ runner.arch }}
           echo ${{ github.workspace }}
-          bash hack/e2e.sh "${HELM_CHART_VERSION}" "${CONTAINER_TAG}" "https://kubean-io.github.io/kubean-helm-chart/" "ghcr.io/kubean-io" "${CONTAINER_TAG}" ${{ runner.name }}
+          bash hack/e2e.sh "${HELM_CHART_VERSION}" "${CONTAINER_TAG}" "https://kubean-io.github.io/kubean-helm-chart/" "ghcr.io/kubean-io" "${CONTAINER_TAG}" ${{ runner.name }} "PR"
