+++ .github/workflows/e2e_schedule.yaml
@@ -87,4 +87,4 @@ jobs:
           CONTAINER_TAG=$(git describe --tags --abbrev=8 --dirty)
           HELM_CHART_VERSION=`echo ${CONTAINER_TAG}|awk -F "-" '{print $1 }'`
           echo ${{ runner.name }}
-          bash hack/sonobouy_e2e.sh "${HELM_CHART_VERSION}" "${CONTAINER_TAG}" "https://kubean-io.github.io/kubean-helm-chart/" "ghcr.io/kubean-io" "${CONTAINER_TAG}" ${{ runner.name }}
+          bash hack/e2e.sh "${HELM_CHART_VERSION}" "${CONTAINER_TAG}" "https://kubean-io.github.io/kubean-helm-chart/" "ghcr.io/kubean-io" "${CONTAINER_TAG}" ${{ runner.name }} "NIGHTLY"
