+++ hack/e2e.sh
@@ -16,6 +16,7 @@ HELM_REPO=${3:-"https://kubean-io.github.io/kubean-helm-chart"}
 IMG_REPO=${4:-"ghcr.io/kubean-io"}
 SPRAY_JOB_VERSION=${5:-latest}
 RUNNER_NAME=${6:-"kubean-actions-runner1"}
+E2E_TYPE=${7:-"PR"}
 EXIT_CODE=0
 
 CLUSTER_PREFIX=kubean-"${IMAGE_VERSION}"-$RANDOM
@@ -39,11 +40,30 @@ chmod +x ./hack/local-up-kindcluster.sh
 chmod +x ./hack/run-e2e.sh
 chmod +x ./hack/run-sonobouy-e2e.sh
 
-utils:runner_ip
-###### e2e logic ########
+DIFF_NIGHTLYE2E=`git show -- '/test/*' | grep nightlye2e || true`
+DIFF_COMPATIBILE=`git show | grep /test/kubean_oscompability_e2e || true`
+
+####### e2e logic ########
 trap utils::clean_up EXIT
 ./hack/local-up-kindcluster.sh "${TARGET_VERSION}" "${IMAGE_VERSION}" "${HELM_REPO}" "${IMG_REPO}" "kindest/node:v1.21.1" "${CLUSTER_PREFIX}"-host
-./hack/run-e2e.sh "${CLUSTER_PREFIX}"-host $SPRAY_JOB_VERSION $vm_ip_addr1
+
+utils:runner_ip
+
+if [ "${E2E_TYPE}" == "PR" ]; then
+    ./hack/run-e2e.sh "${CLUSTER_PREFIX}"-host $SPRAY_JOB_VERSION $vm_ip_addr1
+    # Judge whether to change the nightlye2e case
+    if [[ -n $DIFF_NIGHTLYE2E ]] ; then
+        ./hack/run-sonobouy-e2e.sh "${CLUSTER_PREFIX}"-host $SPRAY_JOB_VERSION $vm_ip_addr1 $vm_ip_addr2
+    fi
+    # Judge whether to change the compitable case
+    if [[ -n $DIFF_COMPATIBILE ]] ; then
+        ./hack/run-os-compitable-e2e.sh "${CLUSTER_PREFIX}"-host $SPRAY_JOB_VERSION $vm_ip_addr1 $vm_ip_addr2
+    fi
+elif [ "${E2E_TYPE}" == "NIGHTLY" ]; then
+    ./hack/run-sonobouy-e2e.sh "${CLUSTER_PREFIX}"-host $SPRAY_JOB_VERSION $vm_ip_addr1 $vm_ip_addr2
+else
+    ./hack/run-os-compitable-e2e.sh "${CLUSTER_PREFIX}"-host $SPRAY_JOB_VERSION $vm_ip_addr1 $vm_ip_addr2
+fi
 
 ret=$?
 if [ ${ret} -ne 0 ]; then
