+++ test/kubeanOps_functions_nightlye2e/e2e-install-cluster/kubeanClusterOps.yml
