+++ test/kubean_add_remove_worker_nightlye2e/add-worker-node/kubeanClusterOps.yml
