+++ test/kubean_add_remove_worker_nightlye2e/e2e-install-1node-cluster-prikey/ssh-auth-secret.yml
