+++ test/kubean_add_remove_worker_nightlye2e/kubean_suite_test.go
@@ -1,4 +1,4 @@
-package kubeanOps_functions_e2e
+package add_worker_e2e
 
 import (
 	"testing"
