+++ test/kubean_add_remove_worker_nightlye2e/remove-worker-node/hosts-conf-cm.yml
