+++ test/kubean_calico_nightlye2e/kubean_network_calico_test.go
@@ -1,4 +1,4 @@
-package kubean_calico_e2e
+package kubean_calico_nightlye2e
 
 import (
 	"context"
