--- test/kubean_oscompitable_e2e/kubean_cluster_install_test.go
@@ -1,12 +1,0 @@
-package kubean_oscompitable_e2e
-
-import (
-	"fmt"
-
-	"github.com/onsi/ginkgo/v2"
-)
-
-var _ = ginkgo.Describe("e2e test cluster operation", func() {
-
-	fmt.Println("yes os")
-})
