+++ test/kubean_sonobouy_nightlye2e/e2e-install-cluster-sonobouy/kubeanClusterOps.yml
